if user_id =0 will show all , else show user of user_id login created

+++ 06ProjectAuthorization/app/Http/Controllers/Admin/UsersController.php
@@ -13,8 +13,11 @@ class UsersController extends Controller
 {
     public function index()
     {
-        $usersList = User::all();
-
+        $userID = Auth::user()->user_id;
+        $usersList = User::where('user_id', $userID)->get();
+        if ($userID == 0) {
+            $usersList = User::where('user_id', '!=', '0')->get();;
+        }
         return view('admin.users.lists', compact('usersList'));
     }
     public function add()
@@ -55,11 +58,13 @@ class UsersController extends Controller
     }
     public function edit(User $user)
     {
+        $this->authorize('users.edit', $user);
         $groupList = Groups::all();
         return view('admin.users.edit', compact('groupList', 'user'));
     }
     public function postEdit(User $user, Request $request)
     {
+        $this->authorize('users.edit', $user);
 
         $rules = [
             'name' => ['required'],
@@ -93,6 +98,8 @@ class UsersController extends Controller
 
     public function delete(User $user)
     {
+        $this->authorize('users.delete', $user);
+
         if ($user->id != Auth::user()->id) {
             User::destroy($user->id);
             return redirect()->route('admin.users.index')->with('msg', 'Xóa người dùng thành công');
